Additional QoL changes

+++ Tasks/git-clone/main.ps1
@@ -125,7 +125,7 @@ function InstallPS7 {
             }
         } -Maximum 5 -Delay 100
         # Need to update the path post install
-        $env:Path = [System.Environment]::GetEnvironmentVariable("Path","Machine") + ";" + [System.Environment]::GetEnvironmentVariable("Path","User")
+        $env:Path = [System.Environment]::GetEnvironmentVariable("Path","Machine") + ";" + [System.Environment]::GetEnvironmentVariable("Path","User") + ";C:\Program Files\PowerShell\7"
         Write-Host "Done Installing PowerShell 7"
     }
     else {
@@ -252,26 +252,48 @@ if (!(Get-Command git -ErrorAction SilentlyContinue)) {
 
         # install git via winget
         Write-Host "Installing git with Install-WinGetPackage"
+        $installExitCode = 123
         $tempOutFile = [System.IO.Path]::GetTempFileName() + ".out.json"
-        $process = Start-Process -FilePath "C:\Program Files\PowerShell\7\pwsh.exe" -ArgumentList "-MTA -Command `"Install-WinGetPackage -Source winget -Id Git.Git | ConvertTo-Json -Depth 10 > $($tempOutFile)`"" -PassThru
-        $handle = $process.Handle # cache process.Handle so ExitCode isn't null when we need it below
-        $process.WaitForExit()
-        $installExitCode = $process.ExitCode
-        $unitResults = Get-Content -Path $tempOutFile
-        Remove-Item -Path $tempOutFile -Force
-        Write-Host "Results:"
-        Write-Host $unitResults
+        $installGitCommand = "Install-WinGetPackage -Source winget -Id Git.Git | ConvertTo-Json -Depth 10 > $($tempOutFile)"
+
+        if (Get-Command pwsh -ErrorAction SilentlyContinue) {
+            if ($PsInstallScope -eq "CurrentUser") {
+                pwsh -Command "$($installGitCommand)"
+            }
+            else {
+                pwsh -MTA -Command "$($installGitCommand)"
+            }
+            $installExitCode = $LASTEXITCODE
+        }
+        else {
+            $process = Start-Process -FilePath "C:\Program Files\PowerShell\7\pwsh.exe" -ArgumentList "-MTA -Command `"$($installGitCommand)`"" -PassThru
+            $handle = $process.Handle # cache process.Handle so ExitCode isn't null when we need it below
+            $process.WaitForExit()
+            $installExitCode = $process.ExitCode
+        }
 
         if ($installExitCode -ne 0) {
-            Write-Error "Failed to install git with Install-WinGetPackage, error code $($installExitCode)"
+            Write-Error "Failed to install Git.Git with Install-WinGetPackage, error code $($installExitCode)"
             # this was the last try, so exit with the install exit code
             exit $installExitCode
         }
 
-        # If there are any errors in the package installation, we need to exit with a non-zero code
-        $unitResultsObject = $unitResults | ConvertFrom-Json
-        if ($unitResultsObject.Status -ne "Ok") {
-            Write-Error "There were errors installing the Git.Git package"
+        # read the output file and write it to the console
+        if (Test-Path -Path $tempOutFile) {
+            $unitResults = Get-Content -Path $tempOutFile
+            Remove-Item -Path $tempOutFile -Force
+            Write-Host "Results:"
+            Write-Host $unitResults
+
+            # If there are any errors in the package installation, we need to exit with a non-zero code
+            $unitResultsObject = $unitResults | ConvertFrom-Json
+            if ($unitResultsObject.Status -ne "Ok") {
+                Write-Error "There were errors installing the Git.Git package"
+                exit 1
+            }
+        }
+        else {
+            Write-Host "Couldn't find output file for Git.Git installation, assuming fail"
             exit 1
         }
 
@@ -302,15 +324,25 @@ if (!(Get-Command git-lfs -ErrorAction SilentlyContinue)) {
 
         # install git-lfs via winget
         Write-Host "Installing git-lfs with Install-WinGetPackage"
+        $installExitCode = 123
         $tempOutFile = [System.IO.Path]::GetTempFileName() + ".out.json"
-        $process = Start-Process -FilePath "C:\Program Files\PowerShell\7\pwsh.exe" -ArgumentList "-MTA -Command `"Install-WinGetPackage -Source winget -Id GitHub.GitLFS | ConvertTo-Json -Depth 10 > $($tempOutFile)`"" -PassThru
-        $handle = $process.Handle # cache process.Handle so ExitCode isn't null when we need it below
-        $process.WaitForExit()
-        $installExitCode = $process.ExitCode
-        $unitResults = Get-Content -Path $tempOutFile
-        Remove-Item -Path $tempOutFile -Force
-        Write-Host "Results:"
-        Write-Host $unitResults
+        $installGitLfsCommand = "Install-WinGetPackage -Source winget -Id GitHub.GitLFS | ConvertTo-Json -Depth 10 > $($tempOutFile)"
+
+        if (Get-Command pwsh -ErrorAction SilentlyContinue) {
+            if ($PsInstallScope -eq "CurrentUser") {
+                pwsh -Command "$($installGitCommand)"
+            }
+            else {
+                pwsh -MTA -Command "$($installGitCommand)"
+            }
+            $installExitCode = $LASTEXITCODE
+        }
+        else {
+            $process = Start-Process -FilePath "C:\Program Files\PowerShell\7\pwsh.exe" -ArgumentList "-MTA -Command `"$($installGitLfsCommand)`"" -PassThru
+            $handle = $process.Handle # cache process.Handle so ExitCode isn't null when we need it below
+            $process.WaitForExit()
+            $installExitCode = $process.ExitCode
+        }
 
         if ($installExitCode -ne 0) {
             Write-Error "Failed to install git-lfs with Install-WinGetPackage, error code $($installExitCode)"
@@ -318,10 +350,22 @@ if (!(Get-Command git-lfs -ErrorAction SilentlyContinue)) {
             exit $installExitCode
         }
 
-        # If there are any errors in the package installation, we need to exit with a non-zero code
-        $unitResultsObject = $unitResults | ConvertFrom-Json
-        if ($unitResultsObject.Status -ne "Ok") {
-            Write-Error "There were errors installing the GitHub.GitLFS package"
+        # read the output file and write it to the console
+        if (Test-Path -Path $tempOutFile) {
+            $unitResults = Get-Content -Path $tempOutFile
+            Remove-Item -Path $tempOutFile -Force
+            Write-Host "Results:"
+            Write-Host $unitResults
+
+            # If there are any errors in the package installation, we need to exit with a non-zero code
+            $unitResultsObject = $unitResults | ConvertFrom-Json
+            if ($unitResultsObject.Status -ne "Ok") {
+                Write-Error "There were errors installing the GitHub.GitLFS package"
+                exit 1
+            }
+        }
+        else {
+            Write-Host "Couldn't find output file for GitHub.GitLFS installation, assuming fail"
             exit 1
         }
